# support inspecting exited frames (faster version)