--- a/src/codegen/irgen/irgenerator.h
+++ b/src/codegen/irgen/irgenerator.h
@@ -158,8 +158,7 @@ public:
     virtual void run(const CFGBlock* block) = 0; // primary entry point
     virtual EndingState getEndingSymbolTable() = 0;
     virtual void doSafePoint(AST_stmt* next_statement) = 0;
-    virtual void addFrameStackmapArgs(PatchpointInfo* pp, AST_stmt* current_stmt,
-                                      std::vector<llvm::Value*>& stackmap_args) = 0;
+    virtual void addFrameStackmapArgs(PatchpointInfo* pp, std::vector<llvm::Value*>& stackmap_args) = 0;
     virtual void addOutgoingExceptionState(ExceptionState exception_state) = 0;
     virtual void setIncomingExceptionState(llvm::SmallVector<ExceptionState, 2> exc_state) = 0;
     virtual llvm::BasicBlock* getCXXExcDest(llvm::BasicBlock* final_dest) = 0;