--- a/src/codegen/unwinding.cpp
+++ b/src/codegen/unwinding.cpp
@@ -353,11 +353,6 @@ public:
     }
 
     AST_stmt* getCurrentStatement() {
-        if (id.type == PythonFrameId::COMPILED) {
-            auto locations = findLocations("!current_stmt");
-            if (locations.size() == 1)
-                return reinterpret_cast<AST_stmt*>(readLocation(locations[0]));
-        }
         assert(getFrameInfo()->stmt);
         return getFrameInfo()->stmt;
     }