--- a/src/runtime/inline/link_forcer.cpp
+++ b/src/runtime/inline/link_forcer.cpp
@@ -72,6 +72,8 @@ void force() {
     FORCE(createPureImaginary);
     FORCE(createSet);
     FORCE(decodeUTF8StringPtr);
+    FORCE(initFrame);
+    FORCE(deinitFrame);
 
     FORCE(getattr);
     FORCE(getattr_capi);