--- a/src/runtime/traceback.h
+++ b/src/runtime/traceback.h
@@ -27,19 +27,26 @@ class GCVisitor;
 extern "C" BoxedClass* traceback_cls;
 class BoxedTraceback : public Box {
 public:
-    Box* tb_next;
     LineInfo line;
-    Box* py_lines;
-    BoxedTraceback(LineInfo line, Box* tb_next) : tb_next(tb_next), line(std::move(line)), py_lines(NULL) {}
+    Box* tb_next;
+    Box* tb_frame;
+
+    BoxedTraceback(LineInfo line, Box* tb_next, Box* tb_frame)
+        : line(std::move(line)), tb_next(tb_next), tb_frame(tb_frame) {
+        if (!tb_frame)
+            this->tb_frame = None;
+        else
+            assert(tb_frame->cls == frame_cls);
+    }
 
     DEFAULT_CLASS(traceback_cls);
 
-    static Box* getLines(Box* b);
+    static Box* lineno(Box* obj, void*);
 
     static void gcHandler(gc::GCVisitor* v, Box* b);
 
     // somewhat equivalent to PyTraceBack_Here
-    static void here(LineInfo lineInfo, Box** tb);
+    static void here(LineInfo lineInfo, Box** tb, Box* frame);
 };
 
 void printTraceback(Box* b);