--- a/src/runtime/types.cpp
+++ b/src/runtime/types.cpp
@@ -105,6 +105,9 @@ void FrameInfo::gcVisit(GCVisitor* visitor) {
     visitor->visit(&exc.type);
     visitor->visit(&exc.value);
     visitor->visit(&frame_obj);
+
+    if (back)
+        ((FrameInfo*)back)->gcVisit(visitor);
 }
 
 // Analogue of PyType_GenericAlloc (default tp_alloc), but should only be used for Pyston classes!