… src/runtime/frame.cpp
… src/runtime/generator.cpp
… src/runtime/traceback.cpp
… test/tests/generator_threads2.py
… test/tests/sys_getframe_exited.py