@kmod 2016-01-29
The 5% regression is worrying, but my thinking is that it's not worth trying to tackle it in isolation.  It seems like the slowdown is coming mostly from a general slowdown (icache contention?) which I think we need to look at as a whole anyway.