@kmod 2016-01-30
One thing I noticed is that a lot of the slowdown (about half on pyxl_bench_10x) comes from setting the current_stmt variable.  Maybe instead of storing it as an `AST_stmt*` pointer, we could store it as an int index into an array of `AST_stmt*`?  This way we could use shorter instructions (8 bytes instead of 14), which might help.